requirements.txt
requests==2.31.0 
pytest==7.4.0 
python-dotenv==1.0.0 
flake8==6.0.0 
